moves sidebar selector to own comp working vis
