+++ app/javascript/containers/home/MainContent.jsx
@@ -1,10 +1,9 @@
 import React from 'react';
 import TagCloud from './components/TagCloud';
 import Results from './components/Results';
+import ContextsSelectorSidebar from './components/ContextsSelectorSidebar';
 
-function classNames(...classes) {
-	return classes.filter(Boolean).join(' ');
-}
+
 
 export default function MainContent(props) {
 	
@@ -13,22 +12,7 @@ export default function MainContent(props) {
 		<>
 			<div className='min-h-0 flex-1 flex overflow-hidden'>
 				{/* //! SIDEBAR */}
-				<nav aria-label='Sidebar' className='hidden md:block md:flex-shrink-0 md:bg-gray-800 md:overflow-y-auto'>
-					<div className='relative w-20 flex flex-col p-3 space-y-3'>
-						{props.contexts.map((c) => (
-							<a
-								key={c.name}
-								href={c.href}
-								className={classNames(
-									c.current ? 'bg-gray-900 text-white' : 'text-gray-400 hover:bg-gray-700',
-									'flex-shrink-0 inline-flex items-center justify-center h-14 w-14 rounded-lg'
-								)}>
-								<span className='sr-only'>{c.name}</span>
-								<c.icon className='h-6 w-6' aria-hidden='true' />
-							</a>
-						))}
-					</div>
-				</nav>
+				<ContextsSelectorSidebar contexts={props.contexts}/>
 
 				{/* //! MAIN CONTENT */}
 				<main className='min-w-0 flex-1 border-t border-gray-200 lg:flex'>
